+++ src/pages/dictionaries/Dictionary.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from "react";
 import DocumentTitle from "react-document-title";
 import { Link, useLocation } from "react-router-dom";
-import DictionaryNotFound from "../../components/DictionaryNotFound";
+import DictionaryNotFound from "./DictionaryNotFound";
 import GoogleAnalytics from "react-ga";
 import PseudoContentButton from "../../components/PseudoContentButton";
 import { IconExternal } from "../../components/Icon";
